Moved tests that need more servers to separate test suite.
