+++ js/client/modules/@arangodb/testsuites/aql.js
@@ -32,7 +32,8 @@ const functionsDocumentation = {
   'shell_client_aql': 'AQL tests in the client',
   'shell_server_aql': 'AQL tests in the server',
   'shell_server_only': 'server specific tests',
-  'shell_client_transaction': 'transaction tests'
+  'shell_client_transaction': 'transaction tests',
+  'shell_client_replication2_recovery': 'replication2 cluster recovery tests'
 };
 const optionsDocumentation = [
   '   - `skipAql`: if set to true the AQL tests are skipped',
@@ -50,6 +51,7 @@ const testPaths = {
   'shell_server_aql': [ tu.pathForTesting('server/aql'), tu.pathForTesting('common/aql') ],
   'shell_client_aql': [ tu.pathForTesting('client/aql'), tu.pathForTesting('common/aql') ],
   'shell_client_transaction': [ tu.pathForTesting('client/shell/transaction')],
+  'shell_client_replication2_recovery': [ tu.pathForTesting('client/shell/transaction/replication2_recovery')],
 };
 
 /// ensure that we have enough db servers in cluster tests
@@ -218,7 +220,26 @@ function shellClientAql (options) {
 
 function shellClientTransaction(options) {
   let testCases = tu.scanTestPaths(testPaths.shell_client_transaction, options);
+  testCases = tu.splitBuckets(options, testCases);
 
+  var opts = ensureServers(options, 3);
+  opts['httpTrustedOrigin'] =  'http://was-erlauben-strunz.it';
+
+  let moreOptions = {
+    "agency.supervision-ok-threshold": "1.5",
+    "agency.supervision-grace-period": "3.0",
+  };
+  let rc = new tu.runLocalInArangoshRunner(opts, 'shell_client_transaction', moreOptions).run(testCases);
+  options.cleanup = options.cleanup && opts.cleanup;
+  return rc;
+}
+
+// //////////////////////////////////////////////////////////////////////////////
+// / @brief TEST: shell_client_replication2_recovery
+// //////////////////////////////////////////////////////////////////////////////
+
+function shellClientReplication2Recovery(options) {
+  let testCases = tu.scanTestPaths(testPaths.shell_client_replication2_recovery, options);
   testCases = tu.splitBuckets(options, testCases);
 
   var opts = ensureServers(options, 5);
@@ -232,7 +253,7 @@ function shellClientTransaction(options) {
     "agency.supervision-ok-threshold": "1.5",
     "agency.supervision-grace-period": "3.0",
   };
-  let rc = new tu.runLocalInArangoshRunner(opts, 'shell_client_transaction', moreOptions).run(testCases);
+  let rc = new tu.runLocalInArangoshRunner(opts, 'shell_client_replication2_recovery', moreOptions).run(testCases);
   options.cleanup = options.cleanup && opts.cleanup;
   return rc;
 }
@@ -246,6 +267,7 @@ exports.setup = function (testFns, defaultFns, opts, fnDocs, optionsDoc, allTest
   testFns['shell_server_aql'] = shellServerAql;
   testFns['shell_server_only'] = shellServerOnly;
   testFns['shell_client_transaction'] = shellClientTransaction;
+  testFns['shell_client_replication2_recovery'] = shellClientReplication2Recovery;
 
   defaultFns.push('shell_api');
   defaultFns.push('shell_client');
@@ -253,6 +275,7 @@ exports.setup = function (testFns, defaultFns, opts, fnDocs, optionsDoc, allTest
   defaultFns.push('shell_client_aql');
   defaultFns.push('shell_server_aql');
   defaultFns.push('shell_client_transaction');
+  defaultFns.push('shell_client_replication2_recovery');
 
   opts['skipAql'] = false;
   opts['skipRanges'] = true;
