+++ tests/Definition/rlog/pr.ps1
@@ -12,5 +12,6 @@ Function global:registerTests()
     registerTest -testname "auto" -index prototype-state-cluster -cluster $true -filter "tests\\js\\common\\shell\\shell-prototype-state-cluster.js"
     registerTest -testname "auto" -index shell-transaction -cluster $true -filter "tests\\js\\client\\shell\\transaction\\shell-transaction.js"
     registerTest -testname "auto" -index shell-transaction-cluster -cluster $true -filter "tests\\js\\client\\shell\\transaction\\shell-transaction-cluster.js"
+    registerTest -testname "auto" -index shell-transaction-replication2-recovery -cluster $true -filter "tests\\js\\client\\shell\\transaction\\replication2_recovery\\shell-transaction-replication2-recovery.js"
     comm
 }
